requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.0
asyncpg==0.29.0
python-ulid==2.0.0
python-dotenv==1.0.0
pydantic>=2.0
pydantic-settings==2.0.3
